Move pagination components and tests into their own subdirectories
